move search in to its own file and add some config values
